+++ apps/backend/src/routes/elections.routes.ts
@@ -15,11 +15,11 @@ import {
 } from '../controllers/elections.controllers.js';
 import { acceptBodyCheck } from '../middlewares/acceptBodyCheck.js';
 import { acceptHeaderCheck } from '../middlewares/acceptHeaderCheck.js';
-import { defaultBallotPaperChecks } from '../middlewares/pathParamChecks/ballotPaperCheck.js';
+import { defaultBallotPaperChecks } from '../middlewares/pathParamChecks/ballotPaperChecks.js';
 import {
   checkElectionNotFrozen,
   defaultElectionChecks,
-} from '../middlewares/pathParamChecks/electionCheck.js';
+} from '../middlewares/pathParamChecks/electionChecks.js';
 import { MimeType } from '../middlewares/utils.js';
 
 export const electionsRouter: Router = Router();
